refactor: extract workspace inventory cleanup
